app/requirements.txt
Django==2.2.6
gunicorn==19.9.0
psycopg2-binary==2.8.3
djangorestframework==3.11.0
Pillow==6.2.2
requests==2.23.0
